+++ source/client/process_impl.h
@@ -21,8 +21,8 @@
 #include "common/filesystem/filesystem_impl.h"
 #include "common/http/context_impl.h"
 #include "common/secret/secret_manager_impl.h"
-#include "common/stats/allocator_impl.h"
 #include "common/stats/thread_local_store.h"
+#include "common/stats_allocator_impl.h"
 #include "common/thread_local/thread_local_impl.h"
 #include "common/upstream/cluster_manager_impl.h"
 #include "common/uri_impl.h"
@@ -36,85 +36,7 @@
 
 #include "extensions/transport_sockets/tls/context_manager_impl.h"
 
-#include "common/stats/allocator_impl.h"
-
 namespace Nighthawk {
-
-class NighthawkCounterImpl : public Envoy::Stats::Counter {
-public:
-  NighthawkCounterImpl(Envoy::Stats::CounterSharedPtr&& inner_counter)
-      : inner_counter_(std::move(inner_counter)) {}
-
-  std::string name() const override { return inner_counter_->name(); };
-  Envoy::Stats::StatName statName() const override { return inner_counter_->statName(); };
-  std::vector<Envoy::Stats::Tag> tags() const override { return inner_counter_->tags(); };
-  std::string tagExtractedName() const override { return inner_counter_->tagExtractedName(); };
-  Envoy::Stats::StatName tagExtractedStatName() const override {
-    return inner_counter_->tagExtractedStatName();
-  };
-  void iterateTagStatNames(const Envoy::Stats::Counter::TagStatNameIterFn& fn) const override {
-    inner_counter_->iterateTagStatNames(fn);
-  };
-  void iterateTags(const Envoy::Stats::Counter::TagIterFn& fn) const override {
-    inner_counter_->iterateTags(fn);
-  };
-  bool used() const override { return inner_counter_->used(); };
-
-  void add(uint64_t amount) override {
-    if (!used()) {
-      per_thread_counters_[std::this_thread::get_id()] = 0;
-    }
-    per_thread_counters_[std::this_thread::get_id()] += amount;
-    inner_counter_->add(amount);
-  };
-  void inc() override { add(1); };
-  uint64_t latch() override { return inner_counter_->latch(); };
-  void reset() override { inner_counter_->reset(); };
-
-  // We return the value we accumulated on this thread, if we have it.
-  // If we don't have it, chances are we are being called from the main
-  // thread. So we return the value from the inner counter, which will be
-  // the merged/global value.
-  uint64_t value() const override {
-    auto it = per_thread_counters_.find(std::this_thread::get_id());
-    if (it != per_thread_counters_.end()) {
-      return it->second;
-    } else {
-      return inner_counter_->value();
-    }
-  };
-
-  Envoy::Stats::SymbolTable& symbolTable() override { return inner_counter_->symbolTable(); }
-  const Envoy::Stats::SymbolTable& constSymbolTable() const override {
-    return inner_counter_->constSymbolTable();
-  }
-
-  // RefcountInterface
-  void incRefCount() override { ++ref_count_; }
-  bool decRefCount() override {
-    ASSERT(ref_count_ >= 1);
-    return --ref_count_ == 0;
-  }
-  uint32_t use_count() const override { return ref_count_; }
-
-private:
-  Envoy::Stats::CounterSharedPtr inner_counter_;
-  std::atomic<uint16_t> ref_count_{0};
-  std::map<std::thread::id, uint64_t> per_thread_counters_;
-};
-
-class NighthawkAllocatorImpl : public Envoy::Stats::AllocatorImpl {
-public:
-  using Envoy::Stats::AllocatorImpl::AllocatorImpl;
-
-  Envoy::Stats::CounterSharedPtr makeCounter(Envoy::Stats::StatName name,
-                                             absl::string_view tag_extracted_name,
-                                             const std::vector<Envoy::Stats::Tag>& tags) override {
-    auto counter = Envoy::Stats::AllocatorImpl::makeCounter(name, tag_extracted_name, tags);
-    return Envoy::Stats::CounterSharedPtr(new NighthawkCounterImpl(std::move(counter)));
-  }
-};
-
 namespace Client {
 
 /**
@@ -142,16 +64,13 @@ private:
   std::vector<StatisticPtr>
   mergeWorkerStatistics(const StatisticFactory& statistic_factory,
                         const std::vector<ClientWorkerPtr>& workers) const;
-
-  std::map<std::string, uint64_t> countersToMap() const;
-
   Envoy::ProcessWide process_wide_;
   Envoy::Thread::ThreadFactoryImplPosix thread_factory_;
   Envoy::Filesystem::InstanceImplPosix file_system_;
   Envoy::Event::TimeSystem& time_system_;
   StoreFactoryImpl store_factory_;
   Envoy::Stats::SymbolTableImpl symbol_table_;
-  NighthawkAllocatorImpl stats_allocator_;
+  StatsAllocatorImpl stats_allocator_;
   Envoy::Stats::ThreadLocalStoreImpl store_root_;
   Envoy::Api::Impl api_;
   Envoy::ThreadLocal::InstanceImpl tls_;
